Udemy/ZTM/Pytorch_DeepLearning/Huggingface_Demo/FoodVision_mini/requirements.txt
gradio==6.0.1
Pillow==12.0.0
torch==2.9.0
torch==2.9.0+cu130
torchvision==0.24.0
torchvision==0.24.0+cu130
